+++ src/main/java/com/creativemd/randomadditions/common/item/items/RandomItemEnergy.java
@@ -4,8 +4,8 @@ import net.minecraft.item.ItemStack;
 
 public interface RandomItemEnergy{
 
-	public abstract int onRecieveEnergy(ItemStack stack, int amount);
+	public abstract float onRecieveEnergy(ItemStack stack, float amount);
 	
-	public abstract int onProduceEnergy(ItemStack stack, int max);
+	public abstract float onProduceEnergy(ItemStack stack, int max);
 
 }
