+++ src/main/java/com/creativemd/randomadditions/common/systems/producer/gui/GuiHeatOMeter.java
@@ -19,7 +19,7 @@ public class GuiHeatOMeter extends GuiPowerOMeter{
 	}
 	
 	@Override
-	public int getPower()
+	public float getPower()
 	{
 		return heat.fuel[index];
 	}
